Move parseAppConfig function to config
