+++ packages/core/src/generate/config/index.ts
@@ -1,1 +1,2 @@
 export * from "./config";
+export * from "./parseAppConfig";
