+++ packages/core/src/generate/generateAppFiles/generatorDefault/generateLinkFileDefault.ts
@@ -1,6 +1,6 @@
 import { TemplateFile, Import } from "../../types";
 import printImport from "../../utils/printImport";
-import { RouteLinkOptions } from "../parseAppConfig";
+import { RouteLinkOptions } from "../../config";
 import { PatternNamedExports } from "../types";
 
 export interface GenerateLinkFileDefaultParams {
