+++ packages/core/src/generate/generateAppFiles/generatorReactRouterV5/generateLinkFileReactRouterV5.ts
@@ -1,6 +1,6 @@
 import { TemplateFile, Import } from "../../types";
 import printImport from "../../utils/printImport";
-import { RouteLinkOptions } from "./../parseAppConfig";
+import { RouteLinkOptions } from "../../config";
 import { PatternNamedExports } from "../types";
 
 export interface GenerateLinkFileReactRouterV5Params {
